Stream should be pushed to pool only if USE-CONNECTION-POOL is T
This fixes a problem when implementing a connection pool outside of Dexador (by
using the fifth return value and passing it back in using :STREAM). Before this
change, the stream was pushed to Dexador's pool on every call. The second time
it was pushed, it was closed (inside PUSH-CONNECTION), resulting in REQUEST
returning an already closed stream as the fifth return value.