Remove STREAM from args on redirect to new server
If a stream was explicitly provided to REQUEST, ignore it on redirects to a new
server. In the fairly likely case that the redirect changes only scheme or
host (but leaves path intact), no progress will be made and the same redirect
will be sent back to Dexador over and over.